Fix for intermittent browser interface failures

+++ common/src/main/scala/com/theguardian/multimedia/archivehunter/common/cmn_models/MediaMetadataMapConverters.scala
@@ -69,7 +69,7 @@ trait MediaMetadataMapConverters {
       value("bit_rate").asInstanceOf[Double],
       value("nb_programs").asInstanceOf[Int],
       value("duration").asInstanceOf[Double],
-      value("size").asInstanceOf[Int],
+      value("size").asInstanceOf[Long],
     )
 
   protected def mappingToMediaMetadata(value:Map[String,AnyVal]) = {
